# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Task and User API
  version: 1.0.0
paths:
...
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request (missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUsers400Response'
  /users/{id}:
    patch:
      summary: Update a user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request (missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchUsersId400Response'

    delete:
      summary: Delete a user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTasksId404Response'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        password:
          type: string
    UserCreate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    GetTasksId404Response:
      type: object
      properties:
        error:
          type: string
          example: "Task not found"
    PostUsers400Response:
      type: object
      properties:
        error:
          type: string
          example: "Email and Password are required"
    PatchUsersId400Response:
      type: object
      properties:
        error:
          type: string
          example: "At least one field (Email or Password) must be provided"